backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
psycopg[binary]==3.2.1
pydantic==2.9.0
python-dotenv==1.0.1

# Providers
openai==1.43.0
google-generativeai==0.7.2
requests==2.32.3
